requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
autopep8==2.0.2
beautifulsoup4==4.12.2
bottle==0.12.25
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
chardet==3.0.4
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==41.0.1
daphne==4.0.0
defusedxml==0.7.1
Django==4.2.3
django-allauth==0.54.0
django-analytical==3.1.0
django-bootstrap4==23.2
django-crispy-forms==2.0
django-environ==0.10.0
django-matomo==0.1.6
django-model-utils==4.3.1
frozenlist==1.3.3
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
honcho==1.1.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==22.10.0
multidict==6.0.4
oauthlib==3.2.2
openai==0.27.8
Pillow==9.5.0
psycopg2==2.9.3
# psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.7.0
pyOpenSSL==23.2.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
sqlparse==0.4.4
tqdm==4.65.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.1
urllib3==2.0.3
whitenoise==6.5.0
yarl==1.9.2
zope.interface==6.0
